requirements.txt
networkx
matplotlib
graphviz
streamlit
st-link-analysis
google-cloud-bigquery
bigframes
pandas